--- ipd/ppp/alembic/versions/577c6bb9f84d_init_dbschema_0_2.py
@@ -1,241 +1,0 @@
-"""init dbschema 0.2
-
-Revision ID: 577c6bb9f84d
-Revises: 
-Create Date: 2024-10-12 23:15:11.247962
-
-"""
-from typing import Sequence, Union
-import sqlmodel
-from alembic import op
-import sqlalchemy as sa
-
-# revision identifiers, used by Alembic.
-revision: str = '577c6bb9f84d'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('dbuser', sa.Column('id', sa.Uuid(), nullable=False),
-                    sa.Column('ispublic', sa.Boolean(), nullable=False),
-                    sa.Column('telemetry', sa.Boolean(), nullable=False),
-                    sa.Column('ghost', sa.Boolean(), nullable=False),
-                    sa.Column('datecreated', sa.DateTime(), nullable=False),
-                    sa.Column('fullname', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('props', sa.JSON(), nullable=True), sa.Column('attrs', sa.JSON(), nullable=True),
-                    sa.PrimaryKeyConstraint('id'), sa.UniqueConstraint('name'))
-    op.create_table('dbFileType', sa.Column('id', sa.Uuid(), nullable=False),
-                    sa.Column('ispublic', sa.Boolean(), nullable=False),
-                    sa.Column('telemetry', sa.Boolean(), nullable=False),
-                    sa.Column('ghost', sa.Boolean(), nullable=False),
-                    sa.Column('datecreated', sa.DateTime(), nullable=False),
-                    sa.Column('props', sa.JSON(), nullable=True), sa.Column('attrs', sa.JSON(), nullable=True),
-                    sa.Column('kind', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('userid', sa.Uuid(), nullable=False),
-                    sa.ForeignKeyConstraint(
-                        ['userid'],
-                        ['dbuser.id'],
-                    ), sa.PrimaryKeyConstraint('id'))
-    op.create_table('dbgroup', sa.Column('id', sa.Uuid(), nullable=False),
-                    sa.Column('ispublic', sa.Boolean(), nullable=False),
-                    sa.Column('telemetry', sa.Boolean(), nullable=False),
-                    sa.Column('ghost', sa.Boolean(), nullable=False),
-                    sa.Column('datecreated', sa.DateTime(), nullable=False),
-                    sa.Column('userid', sa.Uuid(), nullable=False),
-                    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('props', sa.JSON(), nullable=True), sa.Column('attrs', sa.JSON(), nullable=True),
-                    sa.ForeignKeyConstraint(
-                        ['userid'],
-                        ['dbuser.id'],
-                    ), sa.PrimaryKeyConstraint('id'), sa.UniqueConstraint('name'))
-    op.create_table('dbpymolcmd', sa.Column('id', sa.Uuid(), nullable=False),
-                    sa.Column('ispublic', sa.Boolean(), nullable=False),
-                    sa.Column('telemetry', sa.Boolean(), nullable=False),
-                    sa.Column('ghost', sa.Boolean(), nullable=False),
-                    sa.Column('datecreated', sa.DateTime(), nullable=False),
-                    sa.Column('userid', sa.Uuid(), nullable=False),
-                    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('props', sa.JSON(), nullable=True), sa.Column('attrs', sa.JSON(), nullable=True),
-                    sa.Column('desc', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('cmdon', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('cmdoff', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('cmdstart', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('onstart', sa.Boolean(), nullable=False),
-                    sa.Column('ligand', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('sym', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('minchains', sa.Integer(), nullable=False),
-                    sa.Column('maxchains', sa.Integer(), nullable=False),
-                    sa.Column('cmdcheck', sa.Boolean(), nullable=False),
-                    sa.ForeignKeyConstraint(
-                        ['userid'],
-                        ['dbuser.id'],
-                    ), sa.PrimaryKeyConstraint('id'), sa.UniqueConstraint('name'))
-    op.create_table('dbuseruserlink', sa.Column('followerid', sa.Uuid(), nullable=False),
-                    sa.Column('followingid', sa.Uuid(), nullable=False),
-                    sa.ForeignKeyConstraint(
-                        ['followerid'],
-                        ['dbuser.id'],
-                    ), sa.ForeignKeyConstraint(
-                        ['followingid'],
-                        ['dbuser.id'],
-                    ), sa.PrimaryKeyConstraint('followerid', 'followingid'))
-    op.create_table('dbworkflow', sa.Column('id', sa.Uuid(), nullable=False),
-                    sa.Column('ispublic', sa.Boolean(), nullable=False),
-                    sa.Column('telemetry', sa.Boolean(), nullable=False),
-                    sa.Column('ghost', sa.Boolean(), nullable=False),
-                    sa.Column('datecreated', sa.DateTime(), nullable=False),
-                    sa.Column('userid', sa.Uuid(), nullable=False),
-                    sa.Column('desc', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('ordering', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('props', sa.JSON(), nullable=True), sa.Column('attrs', sa.JSON(), nullable=True),
-                    sa.ForeignKeyConstraint(
-                        ['userid'],
-                        ['dbuser.id'],
-                    ), sa.PrimaryKeyConstraint('id'), sa.UniqueConstraint('name'))
-    op.create_table('dbflowstep', sa.Column('id', sa.Uuid(), nullable=False),
-                    sa.Column('ispublic', sa.Boolean(), nullable=False),
-                    sa.Column('telemetry', sa.Boolean(), nullable=False),
-                    sa.Column('ghost', sa.Boolean(), nullable=False),
-                    sa.Column('datecreated', sa.DateTime(), nullable=False),
-                    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('index', sa.Integer(), nullable=False),
-                    sa.Column('props', sa.JSON(), nullable=True), sa.Column('attrs', sa.JSON(), nullable=True),
-                    sa.Column('taskgen', sa.JSON(), nullable=True),
-                    sa.Column('workflowid', sa.Uuid(), nullable=False),
-                    sa.Column('instructions', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.ForeignKeyConstraint(
-                        ['workflowid'],
-                        ['dbworkflow.id'],
-                    ), sa.PrimaryKeyConstraint('id'))
-    op.create_table('dbpoll', sa.Column('id', sa.Uuid(), nullable=False),
-                    sa.Column('ispublic', sa.Boolean(), nullable=False),
-                    sa.Column('telemetry', sa.Boolean(), nullable=False),
-                    sa.Column('ghost', sa.Boolean(), nullable=False),
-                    sa.Column('datecreated', sa.DateTime(), nullable=False),
-                    sa.Column('userid', sa.Uuid(), nullable=False),
-                    sa.Column('desc', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('path', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('cmdstart', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('cmdstop', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('sym', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('ligand', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('props', sa.JSON(), nullable=True), sa.Column('attrs', sa.JSON(), nullable=True),
-                    sa.Column('nchain', sa.Integer(), nullable=False),
-                    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('workflowid', sa.Uuid(), nullable=True),
-                    sa.Column('enddate', sa.DateTime(), nullable=False),
-                    sa.ForeignKeyConstraint(
-                        ['userid'],
-                        ['dbuser.id'],
-                    ), sa.ForeignKeyConstraint(
-                        ['workflowid'],
-                        ['dbworkflow.id'],
-                    ), sa.PrimaryKeyConstraint('id'), sa.UniqueConstraint('name'))
-    op.create_table('dbusergrouplink', sa.Column('userid', sa.Uuid(), nullable=False),
-                    sa.Column('groupid', sa.Uuid(), nullable=False),
-                    sa.ForeignKeyConstraint(
-                        ['groupid'],
-                        ['dbgroup.id'],
-                    ), sa.ForeignKeyConstraint(
-                        ['userid'],
-                        ['dbuser.id'],
-                    ), sa.PrimaryKeyConstraint('userid', 'groupid'))
-    op.create_table('dbpollfile', sa.Column('id', sa.Uuid(), nullable=False),
-                    sa.Column('ispublic', sa.Boolean(), nullable=False),
-                    sa.Column('telemetry', sa.Boolean(), nullable=False),
-                    sa.Column('ghost', sa.Boolean(), nullable=False),
-                    sa.Column('datecreated', sa.DateTime(), nullable=False),
-                    sa.Column('fname', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('tag', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('permafname', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('filecontent', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('props', sa.JSON(), nullable=True), sa.Column('attrs', sa.JSON(), nullable=True),
-                    sa.Column('pollid', sa.Uuid(), nullable=False),
-                    sa.Column('FileTypeid', sa.Uuid(), nullable=True),
-                    sa.Column('parentid', sa.Uuid(), nullable=True),
-                    sa.ForeignKeyConstraint(
-                        ['FileTypeid'],
-                        ['dbFileType.id'],
-                    ), sa.ForeignKeyConstraint(
-                        ['parentid'],
-                        ['dbpollfile.id'],
-                    ), sa.ForeignKeyConstraint(
-                        ['pollid'],
-                        ['dbpoll.id'],
-                    ), sa.PrimaryKeyConstraint('id'))
-    op.create_table('dbpymolcmdflowsteplink', sa.Column('pymolcmdid', sa.Uuid(), nullable=False),
-                    sa.Column('flowstepid', sa.Uuid(), nullable=False),
-                    sa.ForeignKeyConstraint(
-                        ['flowstepid'],
-                        ['dbflowstep.id'],
-                    ), sa.ForeignKeyConstraint(
-                        ['pymolcmdid'],
-                        ['dbpymolcmd.id'],
-                    ), sa.PrimaryKeyConstraint('pymolcmdid', 'flowstepid'))
-    op.create_table('dbreview', sa.Column('id', sa.Uuid(), nullable=False),
-                    sa.Column('ispublic', sa.Boolean(), nullable=False),
-                    sa.Column('telemetry', sa.Boolean(), nullable=False),
-                    sa.Column('ghost', sa.Boolean(), nullable=False),
-                    sa.Column('datecreated', sa.DateTime(), nullable=False),
-                    sa.Column('userid', sa.Uuid(), nullable=False),
-                    sa.Column('grade', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('comment', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('durationsec', sa.Integer(), nullable=False),
-                    sa.Column('props', sa.JSON(), nullable=True), sa.Column('attrs', sa.JSON(), nullable=True),
-                    sa.Column('pollfileid', sa.Uuid(), nullable=False),
-                    sa.Column('pollid', sa.Uuid(), nullable=False),
-                    sa.Column('workflowid', sa.Uuid(), nullable=True),
-                    sa.ForeignKeyConstraint(
-                        ['pollfileid'],
-                        ['dbpollfile.id'],
-                    ), sa.ForeignKeyConstraint(
-                        ['pollid'],
-                        ['dbpoll.id'],
-                    ), sa.ForeignKeyConstraint(
-                        ['userid'],
-                        ['dbuser.id'],
-                    ), sa.ForeignKeyConstraint(
-                        ['workflowid'],
-                        ['dbworkflow.id'],
-                    ), sa.PrimaryKeyConstraint('id'))
-    op.create_table('dbreviewstep', sa.Column('id', sa.Uuid(), nullable=False),
-                    sa.Column('ispublic', sa.Boolean(), nullable=False),
-                    sa.Column('telemetry', sa.Boolean(), nullable=False),
-                    sa.Column('ghost', sa.Boolean(), nullable=False),
-                    sa.Column('datecreated', sa.DateTime(), nullable=False),
-                    sa.Column('props', sa.JSON(), nullable=True), sa.Column('attrs', sa.JSON(), nullable=True),
-                    sa.Column('reviewid', sa.Uuid(), nullable=False),
-                    sa.Column('flowstepid', sa.Uuid(), nullable=False),
-                    sa.Column('task', sa.JSON(), nullable=True),
-                    sa.Column('grade', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('comment', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-                    sa.Column('durationsec', sa.Integer(), nullable=False),
-                    sa.ForeignKeyConstraint(
-                        ['flowstepid'],
-                        ['dbflowstep.id'],
-                    ), sa.ForeignKeyConstraint(
-                        ['reviewid'],
-                        ['dbreview.id'],
-                    ), sa.PrimaryKeyConstraint('id'))
-    # ### end Alembic commands ###
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('dbreviewstep')
-    op.drop_table('dbreview')
-    op.drop_table('dbpymolcmdflowsteplink')
-    op.drop_table('dbpollfile')
-    op.drop_table('dbusergrouplink')
-    op.drop_table('dbpoll')
-    op.drop_table('dbflowstep')
-    op.drop_table('dbworkflow')
-    op.drop_table('dbuseruserlink')
-    op.drop_table('dbpymolcmd')
-    op.drop_table('dbgroup')
-    op.drop_table('dbFileType')
-    op.drop_table('dbuser')
-    # ### end Alembic commands ###
